The erlang compiler erlc sometimes uses {x,1023} as a scratch register.  This causes the beam2jvm compiler to use local variables 0..1024 for x regs, and 1024+ for y registers; resulting in ~double code size because much jvm code uses the wide instruction to address local variables > x255.
This fix rewrites TypeMap to use a Map for x-register allocation in stead of an array [0..1024], thus allowing unused registers (say x7 .. x1022) to not be allocated in the jvm activation record.  The result is much smaller code size, though the compiler may now have become slightly slower.  

For example yaws_config:fload/7 now compiles to ~30000 bytes in stead of ~68000 bytes, which was above max permitted by the JVM.