Special case switch(n) for sets of integers
Elixir has some codegen (for unicode) that creates
huge switches of integers, which all branch to one
or two different targets.  

This code special cases such code (that would
otherwise be lookupswitch) so that we use one or
two bit sets to determine the outcome. This reduces
generated code size significantly for this particular
case.