Special case binary matching BE sub-32-bit ints
This is a common fast-path that now gets code
generated as ESmall. This should give the 
VM a fair change to spot the temp objects.

It would be ideal to be able to code generate it
as int, but then we need to have more control over
the case when such a type is later merged after
a branch.