BIFs erlang:trace/3 and erlang:trace_pattern/3
This implements support for call tracing on
global functions.

The match-spec actions trace, enable_trace and
disable_trace works.

Limitations:

When using it, traced functions called in tail
position become normal calls; so
tracing a loopy function will grow the stack.

Currently, the cost of the injected code
does not go away when tracing
is disabled again (this can be fixed).